[#62] refactor : disable를 적용했던 테스트들 disable 적용 해제
